Change all integer types to Long and all floating point types to Double

+++ src/main/java/com/stripe/model/Card.java
@@ -29,8 +29,8 @@ public class Card extends ExternalAccount {
   String cvcCheck;
   Boolean defaultForCurrency;
   String dynamicLast4;
-  Integer expMonth;
-  Integer expYear;
+  Long expMonth;
+  Long expYear;
   String fingerprint;
   String funding;
   String last4;
